feat(model): Add polymorphic server event models
This commit introduces the data classes required to deserialize JSON messages received from the Zello server.

A polymorphic deserialization strategy is implemented using Jackson annotations on the `ServerCommand` base class. This allows the `command` field in the JSON payload to automatically determine the correct Java object type to create.

Adds the following models:
- ServerCommand (base class)
- LogonResultEvent
- OnChannelStatusEvent
- OnStreamStartEvent
- OnStreamStopEvent
- OnTextMessageEvent
- ChannelUser